Removed tests/classmap.php
Previously, Composer excluded files with filenames ending in
"Test.php" from those it included in generated classmaps. This was
changed in this commit:

https://github.com/composer/composer/commit/8e9659bd8317ad3f623c4d17940380443cf8772c

Because of this change, Composer classmaps can now be used together
with ClassmapFileObjectTestLoader. As such, the existing manually
maintained classmap for Phantestic's test suite is being removed
and its runner modified to use Composer's classmap.